requirements.txt
streamlit==1.50.0
tensorflow==2.16.2
numpy==1.26.4
pillow==10.4.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.3
